Remove MSI; new icon

+++ Build.ps1
@@ -23,9 +23,9 @@ function Create-ArtifactDir
 	mkdir ./artifacts
 }
 
-function Publish-Gzips($version)
+function Publish-Archives($version)
 {
-	$rids = @("linux-x64", "osx-x64")
+	$rids = @("linux-x64", "osx-x64", "win-x64")
 	foreach ($rid in $rids) {
 		& dotnet publish src/SeqCli/SeqCli.csproj -c Release -f $framework -r $rid /p:VersionPrefix=$version /p:ShowLinkerSizeComparison=true
 	    if($LASTEXITCODE -ne 0) { exit 4 }
@@ -33,28 +33,22 @@ function Publish-Gzips($version)
 		# Make sure the archive contains a reasonable root filename
 		mv ./src/SeqCli/bin/Release/$framework/$rid/publish/ ./src/SeqCli/bin/Release/$framework/$rid/seqcli-$version-$rid/
 
-		& ./build/7-zip/7za.exe a -ttar seqcli-$version-$rid.tar ./src/SeqCli/bin/Release/$framework/$rid/seqcli-$version-$rid/
-		if($LASTEXITCODE -ne 0) { exit 5 }
+		if ($rid.StartsWith("win-")) {
+			& ./build/7-zip/7za.exe a -tzip ./artifacts/seqcli-$version-$rid.zip ./src/SeqCli/bin/Release/$framework/$rid/seqcli-$version-$rid/
+			if($LASTEXITCODE -ne 0) { exit 5 }
+		} else {
+			& ./build/7-zip/7za.exe a -ttar seqcli-$version-$rid.tar ./src/SeqCli/bin/Release/$framework/$rid/seqcli-$version-$rid/
+			if($LASTEXITCODE -ne 0) { exit 5 }
 
-		# Back to the original directory name
-		mv ./src/SeqCli/bin/Release/$framework/$rid/seqcli-$version-$rid/ ./src/SeqCli/bin/Release/$framework/$rid/publish/
-		
-		& ./build/7-zip/7za.exe a -tgzip ./artifacts/seqcli-$version-$rid.tar.gz seqcli-$version-$rid.tar
-		if($LASTEXITCODE -ne 0) { exit 6 }
+			# Back to the original directory name
+			mv ./src/SeqCli/bin/Release/$framework/$rid/seqcli-$version-$rid/ ./src/SeqCli/bin/Release/$framework/$rid/publish/
+			
+			& ./build/7-zip/7za.exe a -tgzip ./artifacts/seqcli-$version-$rid.tar.gz seqcli-$version-$rid.tar
+			if($LASTEXITCODE -ne 0) { exit 6 }
 
-		rm seqcli-$version-$rid.tar
+			rm seqcli-$version-$rid.tar
+		}
 	}
-}
-
-function Publish-Msi($version)
-{
-	& dotnet publish ./src/SeqCli/SeqCli.csproj -c Release -f $framework -r win-x64 /p:VersionPrefix=$version /p:ShowLinkerSizeComparison=true
-	if($LASTEXITCODE -ne 0) { exit 7 }
-
-	& msbuild ./setup/SeqCli.Setup/SeqCli.Setup.wixproj /t:Build /p:Configuration=Release /p:Platform=x64 /p:SeqCliVersion=$version /p:BuildProjectReferences=false
-	if($LASTEXITCODE -ne 0) { exit 8 }
-
-	mv ./setup/SeqCli.Setup/bin/Release/seqcli.msi ./artifacts/seqcli-$version.msi
 }
 
 Push-Location $PSScriptRoot
@@ -65,8 +59,7 @@ Write-Output "Building version $version"
 Clean-Output
 Create-ArtifactDir
 Restore-Packages
-Publish-Msi($version)
-Publish-Gzips($version)
+Publish-Archives($version)
 Execute-Tests
 
 Pop-Location
